esercizio d4/w2/m2

+++ m2/w2/d4/W05E02/src/main/java/it/epicode/be/bookingmanagement/config/WebSecurity.java
@@ -46,7 +46,7 @@ public class WebSecurity extends WebSecurityConfigurerAdapter{
     
 	@Override
 	protected void configure(final AuthenticationManagerBuilder auth) throws Exception {
-		Optional<User> authUserObj = userSrv.getUserById((long) 1);
+		Optional<User> authUserObj = userSrv.getUserById(1);
 		User authUser = authUserObj.get();
 		String role = "USER";
 		Set<Role> roles = authUser.getRoles();
