+++ m2/w2/d4/W05E02/src/main/java/it/epicode/be/bookingmanagement/controllers/UserController.java
@@ -6,7 +6,6 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -44,7 +43,7 @@ public class UserController {
 	
 //	@PreAuthorize("hasRole('ADMIN')")
 	@PutMapping("/user/{id}")
-	public ResponseEntity<User> update(@PathVariable Long id, @RequestBody User x) {		
+	public ResponseEntity<User> update(@PathVariable int id, @RequestBody User x) {		
 	    Optional<User> cOpt = userSrv.getUserById(id);		
 	    ResponseEntity<User> check = checkExists(cOpt);
 	    if (check != null) {
